% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_cmi.R
\name{bootstrap_cmi}
\alias{bootstrap_cmi}
\title{Bootstrap conditional mean imputation (CMI) for a censored predictor}
\usage{
bootstrap_cmi(
  analysis_model,
  W,
  Delta,
  Z,
  data,
  est_surv,
  dist = "weibull",
  trapezoidal_rule = FALSE,
  B = 1000
)
}
\arguments{
\item{analysis_model}{A formula input for the final analysis model of interest. Note that}

\item{W}{Column name of observed predictor values (including censored opens).}

\item{Delta}{Column name of censoring indicators. Note that \code{Delta = 0} is interpreted as a censored observation.}

\item{Z}{Column name of additional fully observed covariates.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{est_surv}{A string for which CMI approach to be used: fully-parametric (\code{"FP"}), semiparametric (\code{"SP"}), or nonparametric (\code{"NP"}).}

\item{dist}{(Optional) If \code{est_surv = "FP"}, the assumed distribution for \code{W} in the AFT model, passed to \code{survival::survreg()}. Default is \code{"weibull"}.}

\item{trapezoidal_rule}{A logical input for whether the trapezoidal rule should be used to approximate the integral in the imputed values. Default is \code{FALSE}.}

\item{B}{Number of bootstrap resampling replicates. Default is \code{B = 1000}.}
}
\value{
A dataframe containing the following information for the coefficients of \code{analysis_model}:
\item{Coeff}{Variable name, corresponding to the user-specified model \code{analysis_model}.}
\item{SE}{Bootstrapped standard errors (empirical standard deviation of the bootstrapped estimates).}
\item{LB}{Lowerbound of bootstrapped confidence interval (based on quantiles of bootstrapped estimates).}
\item{UB}{Upperbound of bootstrapped confidence interval (based on quantiles of bootstrapped estimates).}
}
\description{
Bootstrap standard errors / confidence intervals for conditional mean imputation (CMI) for a censored predictor.
}
