% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmi_fp.R
\name{cmi_fp}
\alias{cmi_fp}
\title{Fully parametric conditional mean imputation (CMI) for a censored predictor}
\usage{
cmi_fp(W, Delta, Z, data, fit = NULL, dist = "weibull")
}
\arguments{
\item{W}{Column name of observed predictor values (including censored opens).}

\item{Delta}{Column name of censoring indicators. Note that \code{Delta = 0} is interpreted as a censored observation.}

\item{Z}{Column name of additional fully observed covariates.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{Z}.}

\item{fit}{A \code{survreg} imputation model object modeling \code{W} on \code{Z}. If \code{fit = NULL} (default), the AFT model with only main effects for \code{Z} and assuming a Weibull distribution is fit internally and used.}

\item{dist}{(Optional) Assumed distribution for \code{W} in the AFT model, passed to \code{survival::survreg()}. Default is \code{"weibull"}.}
}
\value{
\item{imputed_data} A copy of \code{data} with added column \code{imp} containing the imputed values.
\item{code} Indicator of algorithm status (\code{TRUE} or \code{FALSE}).
}
\description{
Fully parametric conditional mean imputation (CMI) for a censored predictor using an accelerated failure-time (AFT) model to estimate conditional survival.
}
