% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lm_to_imputed_list.R
\name{fit_lm_to_imputed_list}
\alias{fit_lm_to_imputed_list}
\title{Fit a linear model to a list of imputed datasets}
\usage{
fit_lm_to_imputed_list(imputed_list, formula)
}
\arguments{
\item{imputed_list}{A list with dataframe elements that have been completed via conditional mean imputation, such as by \code{condl_mean_impute_bootstrap()}.}

\item{formula}{A formula object used to fit a linear model to each dataframe in \code{imputed_list}}
}
\value{
A list with the following two elements:
\item{Pooled_Coef}{A vector of the \code{p} average coefficient estimates obtained from fitting \code{formula} to each dataframe in \code{imputed_list}}
\item{Pooled_Var}{A vector of the \code{p} pooled variance estimates, calculated using Rubin's rules}
}
\description{
Takes a list of imputed datasets, as generated by \code{condl_mean_impute_bootstrap()},
and returns a list which summarizes estimation of the \code{p} regression coefficients
}
