% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data with a censored predictor}
\usage{
generate_data(sample_size, setting, cens_type = "right", cens_lod = NULL)
}
\arguments{
\item{sample_size}{Number of independent observations to generate.}

\item{setting}{Choose from six simulation settings in the manuscript: \code{"PH_Light"}, \code{"PH_Heavy"}, \code{"NPH_Light"}, \code{"NPH_Heavy"}, \code{"Light"}, or \code{"Heavy"}.}

\item{cens_type}{Type of censoring (\code{"right"} or \code{"left"}). Default is \code{"right"}.}

\item{cens_lod}{(Optional) Limit of detection. Default is \code{cens_lod = NULL}, which corresponds to random censoring.}
}
\value{
A dataframe with \code{sample_size} rows, each containing one simulated observation comprised of the following columns:
\item{z}{A fully observed standard log-normal covariate.}
\item{x}{True values for the randomly right-censored predictor. This is fully observed and intended for empirical comparisons, though it would not be in practice.}
\item{t}{Observed values for the randomly right-censored predictor.}
\item{d}{Event indicator for \code{t} defined such that \code{d = 1} for uncensored subjects and \code{0} otherwise.}
}
\description{
Generate simulated data with a randomly right-censored predictor
}
