% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data with a censored predictor}
\usage{
generate_data(sample_size, setting)
}
\arguments{
\item{sample_size}{Number of independent observations to generate.}

\item{setting}{Choose from four simulation settings in the manuscript: \code{"PH_Light"}, \code{"PH_Heavy"}, \code{"NPH_Light"}, or \code{"NPH_Heavy"}.}
}
\value{
A dataframe with \code{sample_size} rows, each containing one simulated observation comprised of the following columns:
\item{z}{A fully observed standard log-normal covariate.}
\item{x}{True values for the randomly right-censored predictor. This is fully observed and intended for empirical comparisons, though it would not be in practice.}
\item{t}{Observed values for the randomly right-censored predictor.}
\item{d}{Event indicator for \code{t} defined such that \code{d = 1} for uncensored subjects and \code{0} otherwise.}
}
\description{
Generate simulated data with a randomly right-censored predictor
}
